requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
mysql-connector-python==8.0.33
pydantic==1.10.9
python-dotenv==1.0.0
kafka-python==2.0.2
passlib[bcrypt]==1.7.4
